refactor: AcademyRepository 구조 변경
